go: Refactor for clarity and performance The core of the algorihm has been refactored for improved clarity. Magic numbers were replaced by constants, and the code was reorganised to make it easier to understand.
Previously the rehashing code for blocks were a collision was identified, had a
custom code with a duplicated logic from the core algorithm. This has now being
simplified, resulting in less code and a logic more aligned with what a goasm
implementation would look like.

When it was first implemented, all fields storing state needed for the collision
detection were kept in the digest struct. That was unnecessary, as such data
is only required during block processing, and therefore this commit removes them.

The storage of pre-step compression state has been optimised to use less memory.
The collision detection algorithm only need to access this data for specific
rounds of the SHA1 process which are linked to existing dvs, which at present
is for rounds: 0, 58 and 65.

If more dvs were to be added, targeting different rounds, those would need to
to be stored in memory as well. Note that the original algorithmn stores into
memory the data for the 80 rounds.

Signed-off-by: Paulo Gomes <pjbgf@linux.com>