arm64: Introduce SIMD implementation Introduces a SIMD implementation for the arm64 architecture that is largely based on ARM64 SHA1 and aligns with upstream Go. Extensive comments were added to make it easier to maintain the code going forwards.
The previous non-SIMD implementation was removed, as the SHA1 feature is fairly
well supported across modern ARM64 hardware. Keeping both versions in place
is not worth the additional complexity.

The check for SHA1 feature is based on github.com/klauspost/cpuid/v2. This is a
temporary dependency which will be removed once Go exports the same functionality.

Signed-off-by: Paulo Gomes <pjbgf@linux.com>